+++ src/sys/linux.rs
@@ -4,15 +4,16 @@
 //!
 //! Debug with `systemd-inhibit --list`, `gnome-session-inhibit --list`.
 //!
-//! [ScreenSaver]: https://people.freedesktop.org/~hadess/idle-inhibition-spec/re01.html
-//! [systemd Inhibitor Locks]:(https://www.freedesktop.org/wiki/Software/systemd/inhibit/
+//! [`org.freedesktop.ScreenSaver`]: https://people.freedesktop.org/~hadess/idle-inhibition-spec/re01.html
+//! [systemd Inhibitor Locks]: https://www.freedesktop.org/wiki/Software/systemd/inhibit/
 
-use anyhow::Result;
-use zbus::{blocking::Connection, dbus_proxy};
+use zbus::{blocking::Connection, proxy};
 
 use crate::Options;
 
-#[dbus_proxy(
+pub type Error = zbus::Error;
+
+#[proxy(
     interface = "org.freedesktop.login1.Manager",
     default_service = "org.freedesktop.login1",
     default_path = "/org/freedesktop/login1"
@@ -28,7 +29,7 @@ trait Manager {
     ) -> zbus::Result<zbus::zvariant::OwnedFd>;
 }
 
-#[dbus_proxy(assume_defaults = true)]
+#[proxy(assume_defaults = true)]
 trait ScreenSaver {
     /// Inhibit method
     fn inhibit(&self, application_name: &str, reason_for_inhibit: &str) -> zbus::Result<u32>;
@@ -51,7 +52,7 @@ pub struct KeepActive {
 }
 
 impl KeepActive {
-    pub fn new(options: Options) -> Result<Self> {
+    pub fn new(options: Options) -> Result<Self, zbus::Error> {
         let mut awake = Self {
             options,