+++ src/sys/windows.rs
@@ -7,7 +7,6 @@
 //! [`SetThreadExecutionState`]: https://learn.microsoft.com/en-us/windows/win32/api/winbase/nf-winbase-setthreadexecutionstate
 //! [`PowerSetRequest`]: https://learn.microsoft.com/en-us/windows/win32/api/winbase/nf-winbase-powersetrequest
 
-use anyhow::Result;
 use windows::core::Error as WindowsError;
 use windows::Win32::System::Power::{
     SetThreadExecutionState, ES_AWAYMODE_REQUIRED, ES_CONTINUOUS, ES_DISPLAY_REQUIRED,
@@ -16,13 +15,15 @@ use windows::Win32::System::Power::{
 
 use crate::Options;
 
+pub type Error = WindowsError;
+
 pub struct KeepActive {
     options: Options,
     previous: EXECUTION_STATE,
 }
 
 impl KeepActive {
-    pub fn new(options: Options) -> Result<Self> {
+    pub fn new(options: Options) -> Result<Self, Error> {
         let mut awake = KeepActive {
             options,
             previous: Default::default(),
@@ -31,7 +32,7 @@ impl KeepActive {
         Ok(awake)
     }
 
-    fn set(&mut self) -> Result<(), WindowsError> {
+    fn set(&mut self) -> Result<(), Error> {
         let mut esflags = ES_CONTINUOUS;
 
         if self.options.display {
@@ -49,7 +50,7 @@ impl KeepActive {
         unsafe {
             self.previous = SetThreadExecutionState(esflags);
             if self.previous == EXECUTION_STATE(0) {
-                return Err(WindowsError::from_win32());
+                return Err(WindowsError::from_thread());
             }
 
             Ok(())